+++ CHANGELOG.md
@@ -5,16 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
-
 ## [Unreleased]
 
 ### Fixed
+
+- Fixed bug saving custom fields in Masterdata (receiver, complement, neighborhood, company/building)
 - Not continue to payment when mandatory fields are empty and redirect to form
 
 ## [0.1.2] - 2022-XX-XX
 
 ### Added
+
 - added receiver, complement and suburb with validation and save
 - added google autocomplete view
 - added complement with intl-tel-input
@@ -32,7 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Functionality associated with RICA
 - Functionality associated with Furniture products
 - Functionality associated with Mixed products
+
 ### Fixed
 
 - Fixed issues after test all the cases
-